Move LayoutProperty into styles, add some comments
