+++ src/textual/app.py
@@ -510,7 +510,7 @@ class App(DOMNode):
         # Handle input events that haven't been forwarded
         # If the event has been forwarded it may have bubbled up back to the App
         if isinstance(event, events.Mount):
-            view = View()
+            view = View(id="_root")
             self.register(self, view)
             await self.push_view(view)
             await super().on_event(event)
